net/oic; move oc_buffer.h under header directory.
