+++ net/oic/include/oic/oc_buffer.h
