+++ net/oic/src/messaging/coap/transactions.c
@@ -41,7 +41,7 @@
 #include "oic/port/mynewt/config.h"
 #include "oic/messaging/coap/transactions.h"
 #include "oic/messaging/coap/observe.h"
-#include "api/oc_buffer.h"
+#include "oic/oc_buffer.h"
 
 #ifdef OC_CLIENT
 #include "oic/oc_client_state.h"
