+++ net/oic/src/port/mynewt/lora_adaptor.c
@@ -34,7 +34,6 @@
 #include <node/lora.h>
 
 #include "oic/oc_log.h"
-#include "api/oc_buffer.h"
 #include "oic/port/oc_connectivity.h"
 #include "oic/port/mynewt/adaptor.h"
 #include "oic/port/mynewt/transport.h"
